Hydroponics_Driver/requirements.txt
asn1crypto==0.24.0
azure-iot-device==2.5.1
certifi==2018.8.24
chardet==3.0.4
cryptography==2.6.1
deprecation==2.1.0
entrypoints==0.3
idna==2.6
janus==0.4.0
keyring==17.1.1
keyrings.alt==3.1.1
packaging==20.9
paho-mqtt==1.5.1
pigpio==1.78
pycrypto==2.6.1
PyGObject==3.30.4
pyparsing==2.4.7
PySocks==1.7.1
python-apt==1.8.4.3
pyxdg==0.25
requests==2.21.0
requests-unixsocket==0.2.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
ssh-import-id==5.7
urllib3==1.24.1
